Fix for proper Nomad job scheduling when DockerHub rate limiting has been triggered
By looking at the Nomad documentation at:

https://www.nomadproject.io/docs/drivers/docker#force_pull

it is stated: `If image's tag is latest or omitted, the image will always be pulled regardless of this setting.`

therefore Nomad jobs scheduling will fail as pulling of Docker images
from Docker Hub is rated limited.

The trick is to use a dummy tag (:demo) for all images where latest
tag is used.

Another fix would be to not executer `docker logout` and keeping
the DockerHub credentials but this would be much more insecure.